Implement status tracking.

+++ app/Http/Controllers/CommentController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Comment;
 use Illuminate\Http\Request;
+use Illuminate\Http\Response;
 use Illuminate\Support\Facades\Auth;
 
 class CommentController extends Controller
@@ -11,10 +12,11 @@ class CommentController extends Controller
     private $rules = [
         'content' => ['required', 'string'],
     ];
+
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
     public function index()
     {
@@ -24,7 +26,7 @@ class CommentController extends Controller
     /**
      * Show the form for creating a new resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
     public function create()
     {
@@ -34,8 +36,8 @@ class CommentController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
+     * @param Request $request
+     * @return Response
      */
     public function store(Request $request, $post_id)
     {
@@ -44,13 +46,14 @@ class CommentController extends Controller
         $data['user_id'] = Auth::id();
         $data['post_id'] = $post_id;
         $comment = Comment::create($data);
+        return redirect()->route('post.list');
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  \App\Models\Comment  $comment
-     * @return \Illuminate\Http\Response
+     * @param Comment $comment
+     * @return Response
      */
     public function show(Comment $comment)
     {
@@ -60,8 +63,8 @@ class CommentController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  \App\Models\Comment  $comment
-     * @return \Illuminate\Http\Response
+     * @param Comment $comment
+     * @return Response
      */
     public function edit(Comment $comment)
     {
@@ -71,9 +74,9 @@ class CommentController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\Comment  $comment
-     * @return \Illuminate\Http\Response
+     * @param Request $request
+     * @param Comment $comment
+     * @return Response
      */
     public function update(Request $request, $id)
     {
@@ -81,16 +84,20 @@ class CommentController extends Controller
         $comment = Comment::find($id);
         $comment->content = $request->content;
         $comment->save();
+        return redirect()->route('post.list');
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Models\Comment  $comment
-     * @return \Illuminate\Http\Response
+     * @param Comment $comment
+     * @return Response
      */
     public function destroy($id)
     {
-        Comment::destroy($id);
+        if (Auth::User()->role == 'admin' || Auth::id() == Comment::find($id)->user->id) {
+            Comment::destroy($id);
+        }
+        return redirect()->route('post.list');
     }
 }
